// src/app/game/models/player.model.ts
export enum PlayerAnimations {
  IDLE = 'idle',
  RUN = 'run',
  JUMP = 'jump',
  FALL = 'fall',
  KICK = 'kick',
  DEATH = 'death',
  RIFLE_IDLE = 'rifle_idle',
  RIFLE_RUN = 'rifle_run',
  RIFLE_JUMP = 'rifle_jump',
}

export enum PlayerActionStates {
  IDLE = 'idle',
  RUN = 'run',
  JUMP = 'jump',
  FALL = 'fall',
  KICK = 'kick',
  DEATH = 'death',
}

export enum PlayerWieldingStates {
  NOTHING = 'nothing',
  RIFLE = 'rifle'
}